requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.5.0; python_version >= '3.7'
cffi==1.15.0
cryptography==36.0.1; python_version >= '3.6'
django-filter==21.1
django-rest-knox==4.2.0
django==4.0.3
djangorestframework==3.13.1
pycparser==2.21
pytz==2021.3
sqlparse==0.4.2; python_version >= '3.5'
tzdata==2021.5; sys_platform == 'win32'
